Overhaul
Changed quite a bit. There was a bug with the difficulty levels where the boards generated were impossible to solve for easy and medium levels.
I changed the rationale to one in which the board is an instance of the Solver object and the methodology for building it is correct.
Also added a JFrame with a button on it to display the correct answer once the user is sick of trying.