Added visual element
added a new class called board. This class is Solver's parent class and its purpose is to create a JTable to host the sudoku board.

This is the first step towards making this a more interactive app and now we have a visual representation of the board generated by the computer.

next step is to make the instantiation also happen on the JPanel and not the terminal.